Initial Flask server app
This adds enough of a skeleton to have a Flask app running and listing
available resources for GET /.  It also adds an example config file to
use with the server.  For now it can be run as 'enamel --config-file
example.conf'.  The servers resource is not fully implemented but is in
place to show that GET / works.