Integrate gabbi tests by de-globalizing the flask app
Gabbi prefers an app factory to use to get the WSGI app that it
will intercept for testing. Once we have this factory it then makes
sense to not make 'app' a module level variable and instead pass it
around to configuration functions.

The upshot of this is that we now have a single function which adds
route-function pairs and where we do need to get 'app' we can request
it from the flask module property 'current_app'. If you are of a mind
that global stuff is bad and that centralized routing is good, this
is all a win.

The gabbi test in basic.yaml is expecting a response formatted as JSON
Home. Since it is not currently getting that it is marked to xfail.