Adjust service startup to allow keystonemiddleware
Some minor hoop jumping is required to use keystonemiddleware.auth_token
without paste: It's necessary to pass it a pointer to the oslo-config-
using project (in this case enamel) so it can find its own config.

At the same time it is useful to be able to turn off keystone for
functional tests or maybe some other auth strategy is desired. To
that end an 'auth_strategy' config option has been added. The gabbi
config fixtures have been extended to allow setting this to none.

Some potential issues with this setup:

* As is, all routes require auth. With things like / (which will return
  the json-home info) or /schemas it might be nice to be open.
* Currently the loading and applying of middleware is a crufty
  process in create_app rather than something declarative (like what
  paste would provide). It is not clear this is a problem. Very
  little middleware will be desired and pretty much all of it will
  be required and that's the way we want it to remain.