Make config more effective for auth_middleware
This is not yet fully working because of bug #1540022 in
keystonemiddleware, however once fully working it will allow passing
the active and already loaded configuration into the middleware
in a straightforward fashion from both the test and non-test
environments and get rid of the call to the filter_factory (which
was redundant).

And the above bug is addressed the tests still pass and the server
still runs but it won't pay attention to the config values being set
in the example.conf file (when passed as a --config-file arg). For
the time being this isn't an issue as we aren't really doing anything
with auth. In live tests disable keystone setting auth_strategy to
something other than "keystone".

Additionally, logging is configured to ensure logging is visible
and logging and other configuration are centralized in a
prepare_service method so it can be reused by functional tests,
if necessary.