Initial devstack plugin for enamel
As we only have the api service at this point, this doesn't do much
other than set up config and start the api service. Several things
to note:

* an app.wsgi has been created so that mod_wsgi or other wsgi
  servers can host the api server
* The server can run under mod_wsgi or with its own entry point. If
  the latter is used a port (defaulting to 5050 is used). If
  mod_wsgi is used, then the api is hosted on port 80 with a prefix
  of the service type.
* The service type is currently 'enamel'. This will be wrong,
  eventually.
* There is a commented out line to enamel-dbsync, which is waiting
  on a branch.

The idea here is that from this starting point we can start adding
on worker services and eventually multi-node configurations (by
disabling enamel-api and enabling enamel-worker on a different
node).

Before this is merged it can be tested by adding to local.conf:

  enable_plugin gnocchi https://github.com/cdent/enamel devstack-plugin

The defaults (in devstack/settings) make it so that things will
"just work". If ENABLE_HTTPD_MOD_WSGI_SERVICES is True the server
will be at <your ip>/enamel, if False it will be localhost:5050/.