Initial microversion header handling
This puts in place the basic extraction of a Version object from
an Openstack-enamel-api-version header and saves it to the request
context `g` used by flask. Responses have a vary header extended to
include the openstack-* header and that same header with a value of
the selected version is also sent.

A Version is namedtuple with a major and minor element. Tuples like
this have built in comparability so we don't have to write that code
ourselves.

A VERSIONS list contains a list of all stringified valid
microversions of this API. The data currently in there is just stuff
I made up.

There are min_version_string and max_version_string methods which
can be used to declare the window of supported VERSIONS. The idea
being that this _could_ be different than the start and end of
VERSIONS.

As specified in the guidelines, if not version header is sent, the
version used is the min_version. If 'latest' is sent, you get
max_version.

If you send a version which is between min and max but not in
VERSIONS you get an error. This provides a safetly valve for
declaring "crap we made a bad version and we absolutely must
_not_ use it".

Bad version choices, or sending a string that is not a number will
cause an error. The error at the moment is a ValueError but that
will be translated to a 406 when we have proper error response
handling.

This doesn't do it yet, but the theory here is that with the
availability of `g.version` it will be possible to both decorate
methods and do conditionals to say "this will only work for version
window blah" much in the same that nova does.

To get the unit tests working there's some very rudimentary base
class stuff added.