Set api.version.VERSIONS to real value
The original value was for the sake of testing. Now testing is
handled by monkey patching VERSIONS from the tests. In the process
of doing this some issues with how the Version class was structured
were discovered.

Now the Version class is used to simply represent the version itself
and the constraints thereof. Code is extracted to module level
functions where it cannot be subclassed (we don't want it to be).
In the process cycles and dependencies in import handling are
avoided (making the monkey patching cleaner).