Implement an error handler
This makes it possible to raise HTTP<some status> anywhere in the
code and for it to be properly handled by the Flask app. The
response status will be set and the body will take on the form of an
OpenStack errors object.

This implementation uses my own httpexceptor module. We don't have
to use it by find it simple to use and it is fairly reasonable with
regard to headers it will mess with depending on the exception. For
example for most exception the argument passed is just a message
that goes in the error, whereas for 302 it is expected to be a full
URL for redirection. We can change to something else easily enough.

Version handling now properly raises the expected 406 when given a
bad microversion.