Add request-id handling
This adds a simple uuid4 to every response and makes that uuid
available throughout the request, for logging and related purposes.

During testing it became clear that the error handling would not
work for 404s that Flask itself was raising so those are repackaged
to get them into the error handling flow.