Add very basic integration testing harness
Add the foundation for a gabbi-run-based integration test harness
for enamel. The idea is that from one YAML file we can, eventually,
test the whole service (and its collaborators) via the API.

This involves a few moving parts:

* A working devstack plugin (there was a typo in a conditional
  preventing the ENAMEL_SERVICE_URL from being set).
* A post_test_hook.sh which can be used either as a manual script
  to call the tests (once a devstack has been stacked) or as a gate
  hook to be called from project/jenkins config.
* A gabbi yaml file that for now just checks the service is there.
* A tox 'integration' target called by the hook script.

The hook will need some additions to properly report errors and
logging information once this stuff is in the gate. For now we can
do things by hand if we feel inclined.