Switch to using modern microversion headers
The proposed new styleeee is "OpenStack-API-Version: <type> <version>"

This change implements that and updates the tests accordingly.

The proposal has not yet been approved but has support from the cores
of the API-wg and Nova and Ironic are okay with it as is. Manila
is okay as long as there is some grandfathering.